fix: pass CLIENT_FOUND_ROWS
This changes the behaviour of the affected count returned for mutations (UPDATE/INSERT etc). It makes MySQL returns the FOUND rows instead of the AFFECTED rows. Without this, MySQL returns 0 affected row when performing a noop update. This breaks the QueryEngine when performing nested connects which expect the update to have had happened.